Adjust first step to use subtraction of gradients
Using Sobel operator to obtain the second derivative both in x and y,
the resulting image is obtained from the difference of the x-gradient with the y-gradient.
For the snapshot, only this result is processed as a step.

Additionally, removed redundant step of conversion of binary image to color.

Also separated the painting and description of the barcode to a new "step", as
well as strengthened said step against failures in analysis of the barcode.